moved Listing class and its corresponding test class to polybazaar package
