+++ app/src/androidTest/java/ch/epfl/polybazaar/ListingTest.java
@@ -1,4 +1,4 @@
-package ch.epfl.sdp;
+package ch.epfl.polybazaar;
 
 import org.junit.Test;
